Refactor prompt engineering and add prompts for recommendation and cause steps.
