+++ repeated_calls/prompt_engineering/templates/recommendation_system.j2
@@ -1,0 +1,5 @@
+Your job is to determine if the customer is eligible for a discount based on their customer lifetime value (CLV).
+To do this, you will be provided with a customer ID and a product ID, and you must identify the associated discounts.
+Write a message to the user with the customer ID, product ID, CLV value, discount percentage, and duration in months.
+If the customer is not eligible for a discount, write a message to the user indicating that.
+Assure the customer that they are valued and that the company is working to improve their experience.
