+++ repeated_calls/prompt_engineering/templates/recommendation_user.j2
@@ -1,0 +1,5 @@
+Customer ID: {{ cause_result.customer_id }}
+Product ID: {{ cause_result.product_id }}
+CLV Value: {{ clv_value }}
+Discount: {{ discount_percentage }}%
+Duration: {{ discount_duration_months }} months
